Replace POSIX message queue with SystemV semaphore as this is more widely available and also appears less error-prone.
