+++ ckernel/base_kernel.py
@@ -12,7 +12,8 @@ from ipykernel.ipkernel import IPythonKernel
 
 import ckernel
 
-from .util import STDERR, STDOUT, Stream, Trigger
+from .util import STDERR, STDOUT, Stream
+from .trigger import AbstractTrigger
 
 
 class BaseKernel(IPythonKernel):
@@ -102,7 +103,7 @@ class BaseKernel(IPythonKernel):
         return self.stream_data(STDERR, reader, end="")
 
     def write_input(
-        self, writer: asyncio.StreamWriter, trigger: Trigger, prompt: str = ""
+        self, writer: asyncio.StreamWriter, trigger: AbstractTrigger, prompt: str = ""
     ):
         """Get input and send to writer. Wait for input request from trigger"""
         while True:
